WIP: Index crates also using name and path
Well, it's nice to house the data also using crate names, not only
paths to json files with analysis data, but since global symbol id
calculation takes only name into account, I think this will probably
have to be changed somehow, ideally on the compiler front? To at least
include version data and possibly path to those actual crates, not
where they are referenced (who needs that anyway?)

Because in practice, I think we need to guarantee that crate id maps
the same way as it does when calculating the global id - otherwise we
might get a conflict.